refactor: move shared components to helpers folder
