+++ apps/admin/components/chat/rooms/participants.tsx
@@ -13,7 +13,7 @@ import { PlusIcon } from 'lucide-react';
 import { User } from '@/lib/models/User';
 import { Profile } from '@/components/chat/helpers/profile';
 import { ParticipantsLogs } from '@/lib/models/chat';
-import { DeleteAlert } from '@/components/chat/helpers/delete';
+import { DeleteAlert } from '@/components/helpers/delete';
 import { usePathname, useRouter } from 'next/navigation';
 
 type RoomResponse = Awaited<ReturnType<typeof getRoomById>>;
